Add --dry-run

+++ Python/tca_scaffold/tca_scaffold.py
@@ -3,12 +3,17 @@ from jinja2 import Environment, FileSystemLoader
 import click
 import sys
 
-def process_template(script_dir, two_files, force_overwrite, feature_name):
 
-    os.chdir(script_dir)
+def make_abs_path(rel_path):
+    script_dir = os.path.dirname(__file__)
+    return os.path.normpath(os.path.join(script_dir, rel_path))
+
+def process_template(script_dir, two_files, force_overwrite, dry_run, feature_name):
+
+    templates_path = make_abs_path('templates')
 
     # Define the template directory
-    file_loader = FileSystemLoader('templates')
+    file_loader = FileSystemLoader(templates_path)
 
     # Create the environment
     env = Environment(loader=file_loader)
@@ -41,29 +46,34 @@ def process_template(script_dir, two_files, force_overwrite, feature_name):
 
     # Open the file for writing
     filename = f"{feature_name}View.swift"
-    print(f"Creating stub for {feature_name}")
 
-    if os.path.isdir(filename):
-        print(f'Output destination "{filename}" already exists and is a directory, refusing to overwrite.')
-        print()
-        sys.exit()
+    console_prefix = "(DRY RUN:) " if dry_run else ""
 
-    if not force_overwrite and os.path.isfile(filename):
-        print(f'Output destination "{filename}" already exists, refusing to overwrite. Use --force-overwrite to ignore existing files.')
-        print()
-        sys.exit()
+    print(f"{console_prefix}Creating stub for {feature_name}")
 
-    with open(filename, 'w') as f:
-        f.write(stub_contents)
+    if not dry_run:
+        if os.path.isdir(filename):
+            print(f'A directory named "{filename}" already exists, refusing to overwrite.')
+            print()
+            sys.exit(1)
+
+        if not force_overwrite and os.path.isfile(filename):
+            print(f'A file named "{filename}" already exists, refusing to overwrite. Use --force-overwrite to suppress this error.')
+            print()
+            sys.exit(1)
+
+        with open(filename, 'w') as f:
+            f.write(stub_contents)
 
     # output = view_feature_template.render(substitutions)
     # print(output)
 
 @click.command(no_args_is_help=True)
-@click.option('--two-files', is_flag=True, help='Put view and reducer into separate files (UNIMPLEMENTED)')
-@click.option('--force-overwrite', is_flag=True, help='Force overwriting any existing files')
+@click.option('--two-files', is_flag=True, help="Put view and reducer into separate files.")
+@click.option('--force-overwrite', is_flag=True, help="Force overwriting any existing files.")
+@click.option('--dry-run', is_flag=True, help="Don't generate files, just preview any actions")
 @click.argument('feature_names', nargs=-1)
-def start(two_files, force_overwrite, feature_names):
+def start(two_files, force_overwrite, dry_run, feature_names):
     # print(f"Args: {feature_names}")
 
     if len(feature_names) < 1:
@@ -78,7 +88,7 @@ def start(two_files, force_overwrite, feature_names):
     script_dir = os.path.abspath(os.path.dirname(__file__))
 
     for feature_name in feature_names:
-        process_template(script_dir, two_files, force_overwrite, feature_name)
+        process_template(script_dir, two_files, force_overwrite, dry_run, feature_name)
 
     print()
     print("Done")
